--- internal/classifier/search.go
@@ -1,7 +1,0 @@
-package classifier
-
-import "github.com/bitmagnet-io/bitmagnet/internal/database/search"
-
-type searchAction struct {
-	search search.Search
-}
