+++ internal/model/content_source.go
@@ -1,0 +1,5 @@
+package model
+
+const SourceTmdb = "tmdb"
+const SourceImdb = "imdb"
+const SourceTvdb = "tvdb"
